Move table name specifications to constants
The next step is to deploy this on production. However I don't want to
write to the main table just yet (to keep things completely separate
between lazer and legacy scores), to allow for testing out performance
and structural changes without worrying about locking the whole table.

This will allow me to create different named temporary tables to test
again. Haven't bothered with `solo_scores_process_history` just yet as
it is out of scope for the initial testing.