Add command to migrate playlist scores from `multiplayer_scores` to `solo_scores`
This is to be used to preserve ongoing featured artist / spotlight
room's scores as we migrate scores to the new table.

It means we can continue to move forward without worrying about breaking
things, basically.

It's made to only run for playlists, and not migrate *all* scores, but
could be expanded in the future if we want to have more coverage.