Also refresh beatmap store every 60 seconds
This is just the simplest way to ensure we have an up-to-date "ranked"
state for beatmaps as they change states.

We'll want to change this in the future.

Note that this is the same code I removed from the total processor. It's
not as bad to run here because at most there's one beatmap lookup per
user score, rather than thousands (due to looking up every score set by
every user ever).