Reformulate "difficulty reduction" check again
I guarantee this is going to get re-used for "hush-hush" medals whenever
that happens, so it feels better to have a single helper method for
those. Made it medal-specific too by not having it be an extension,
therefore requiring the class name to be used, and as such clearly
binding it to medals specifically.