Constrain query to preserved partition
I'm not sure this is going to do much since we're still preserving more
than we should (I think?) but at least this should eliminate rotated
partitions from the query. Can't be worse.

Notably this relies on https://github.com/ppy/osu-web/pull/10946 to work
correctly (the initial `preserve` value is important, as the query
checking for medal completion already presumes the row to be in a
"final" state for querying).