Unapply non-buffered performance optimisation
ProxySQL doesn't work great with this. If there's a pause (or sometimes
even if there isn't) ie. waiting for replication to catch up, the
connection may be dropped causing the reading to fail and abort the
whole process.

I'm PRing this separately because I may bring this back in the near
future, depending on how things go. Right now we're bottlenecked on
writes, not reads, so it's not a huge issue.