Fix faulty production check in database test setup
The check *works*, but it dies a little harder than it should. Selecting
`*` when the type param in the `QueryFirstOrDefault<>()` invocation is
`int?` is going to die, because dapper can't map 2 columns to one
`int?`.

By chance this would work correctly on *non*-production databases,
because if the row isn't there, then dapper just does the default part
of `QueryFirstOrDefault<>()` which is `null`.