Split domain to which LIO requests are directed to envvar
On several occasions, I've tried to test something involving LIO
full-stack and forgot that the LIO helper hardcodes production.
Thankfully I never did damage due to the interop secret check, but it's
a bit bone-chilling every time. Additionally I believe this situation
can't be correct as far as staging is involved, either.

Same as the shared interop secret, this envvar is optional in debug (and
it missing causes LIO to become a no-op), and required with no default
in release.