Fix mania keymode user stats processor completely bungling convert treatment
The TODO about key mods is one thing, the flagrantly idiotic
reinterpretation of `diff_size` as key count always regardless of
original beatmap ruleset is another.

This is not a complete solution because the "initial playcount
estimation" is pretty much infeasible to include converts in, but it is
at least supposed to stop the bleeding on some complete idiocy.

And yeah this will likely also mean recomputation of these stats game
wide. Probably.

I would ask who bungled the previous implementation *that* badly, but I
know I would need to look in the mirror to answer that question.