Reduce web request timeouts
After deploying the latest changes whose aim is to emit high scoreboard
rank events for lazer leaderboards, many systems started having a
conniption around 00:00 UTC.

The likely explanation is that some job is running every day at midnight
UTC, which causes `osu-web` to respond to API requests made by
`osu-queue-score-statistics`  for scores with some delay, which -
because of `osu-queue-score-statistics` performing the request inside a
transaction - probably is bringing down everything else.

Performing web requests with a transaction held is suboptimal to say the
least, but let's start with decreasing the timeout for now given that
the rank achieved stuff is generally non-critical to the functioning of
the game.

This timeout will also apply to LIO requests, which I think is probably
fine - LIO requests will retry up to three times anyway.